WIP move JS classes into lib/modules
