+++ lib/add-contributor.js
@@ -1,7 +1,7 @@
 module.exports = addContributor;
 
 const getUserDetails = require("./get-user-details");
-const ContentFiles = require("./content-files");
+const ContentFiles = require("./modules/content-files");
 
 async function addContributor({
   context,
