+++ lib/get-config.js
@@ -1,7 +1,7 @@
 module.exports = getConfig;
 
-const Config = require("./config");
-const { ResourceNotFoundError } = require("./errors");
+const Config = require("./modules/config");
+const { ResourceNotFoundError } = require("./modules/errors");
 
 async function getConfig(repository) {
   const config = new Config(repository);
