+++ lib/get-user-details.js
@@ -1,6 +1,6 @@
 module.exports = getUserDetails;
 
-const { UserNotFoundError } = require("./errors");
+const { UserNotFoundError } = require("./modules/errors");
 
 async function getUserDetails({ octokit, username }) {
   // TODO: optimzation, if commenting user is the user we're adding we can avoid an api call
