+++ lib/modules/config.js
@@ -35,7 +35,8 @@ class Config {
       content: rawOptionsFileContent,
       sha,
     } = await this.repository.getFile(ALL_CONTRIBUTORS_RC);
-    this.originalOptionsSha = sha;
+    this.fileSha = sha;
+
     try {
       this.options = JSON.parse(rawOptionsFileContent);
 
@@ -86,7 +87,7 @@ class Config {
   }
 
   getOriginalSha() {
-    return this.originalOptionsSha;
+    return this.fileSha;
   }
 
   async addContributor({ login, contributions, name, avatar_url, profile }) {
