+++ lib/modules/content-files.js
