+++ lib/modules/errors.js
