+++ lib/setup-repository.js
@@ -1,7 +1,7 @@
 module.exports = setupRepository;
 
-const Repository = require("./repository");
-const { BranchNotFoundError } = require("./errors");
+const Repository = require("./modules/repository");
+const { BranchNotFoundError } = require("./modules/errors");
 
 async function setupRepository({ context, branchName }) {
   const repository = new Repository(context);
